api/requirements.txt
fastapi[standard]
sqlmodel
requests
python-jose[cryptography]
python-multipart
psycopg2-binary
